Add files via upload
ColorReducer.R will reduce all the image colors to the 5 target chosen colors by applying K-means clustering and assigning the original image color to the nearest (Euclidean distance) centroid of the target color. The result of this will be the modified image formed after re-assigning the clustered image colors and replacing it with the target colors.